+++ app/Settings/Admin/OptionsPage.php
@@ -192,10 +192,10 @@ class OptionsPage implements Bootable {
 	public function template() { ?>
 
 		<div class="wrap">
-			<h1 class="wp-heading-inline"><?= esc_html( $this->label ) ?></h1>
+			<h1 class="wp-heading-inline"><?php esc_html( $this->label ) ?></h1>
 
 			<?php if ( current_user_can( 'customize' ) ) : ?>
-				<a href="<?= esc_url( wp_customize_url() ) ?>" class="hide-if-no-js page-title-action"><?php esc_html_e( 'Customize', 'amicable' ) ?></a>
+				<a href="<?php esc_url( wp_customize_url() ) ?>" class="hide-if-no-js page-title-action"><?php esc_html_e( 'Customize', 'amicable' ) ?></a>
 			<?php endif ?>
 
 			<div class="wp-filter">
@@ -231,7 +231,7 @@ class OptionsPage implements Bootable {
 					$url = remove_query_arg( 'view', $url );
 				endif; ?>
 
-				<li class="<?= sanitize_html_class( $view->name() ) ?>">
+				<li class="<?php sanitize_html_class( $view->name() ) ?>">
 					<?php printf(
 						'<a href="%s"%s>%s</a>',
 						esc_url( $url ),