+++ app/Settings/Admin/Views/General.php
@@ -86,7 +86,7 @@ class General extends View {
 	 * @param  array  $input
 	 * @return array
 	 */
-	function validateSettings( $settings ) {
+	public function validateSettings( $settings ) {
 
 		// Checkboxes.
 		$settings['disable_emoji']    = ! empty( $settings['disable_emoji']    );
@@ -224,7 +224,7 @@ class General extends View {
 			'name'              => 'amicable_settings[error_page]',
 			'show_option_none'  => '-',
 			'option_none_value' => 0,
-			'selected'          => Options::get( 'error_page' ),
+			'selected'          => Options::get( 'error_page' ), // phpcs:ignore
 			'post_status'       => [ 'private' ],
 			'echo'              => false
 		] ); ?>
@@ -233,7 +233,7 @@ class General extends View {
 			<label>
 				<?php if ( $dropdown ) : ?>
 
-					<?= $dropdown ?>
+					<?php echo $dropdown // phpcs:ignore ?>
 
 				<?php else : ?>
 
@@ -243,7 +243,7 @@ class General extends View {
 
 					<?php if ( current_user_can( 'publish_pages' ) ) : ?>
 
-						<a href="<?= esc_url( add_query_arg( 'post_type', 'page', admin_url( 'post-new.php' ) ) ) ?>"><?php esc_html_e( 'Add New Page', 'amicable' ) ?></a>
+						<a href="<?php esc_url( add_query_arg( 'post_type', 'page', admin_url( 'post-new.php' ) ) ) ?>"><?php esc_html_e( 'Add New Page', 'amicable' ) ?></a>
 
 					<?php endif ?>