+++ webpack.mix.js
@@ -68,7 +68,9 @@ mix.version();
  * @link https://laravel.com/docs/5.6/mix#working-with-scripts
  */
 mix.js( `${devPath}/js/app.js`, 'js' )
-	.js( `${devPath}/js/navigation.js`, 'js' );
+	.js( `${devPath}/js/navigation.js`, 'js' )
+	.js( `${devPath}/js/customize-controls.js`, 'js' )
+	.js( `${devPath}/js/customize-preview.js`,  'js' );
 
 /*
  * Compile CSS. Mix supports Sass, Less, Stylus, and plain CSS, and has functions
@@ -80,4 +82,6 @@ mix.js( `${devPath}/js/app.js`, 'js' )
  */
 
 // Compile SASS/CSS.
-mix.sass( `${devPath}/scss/screen.scss`, 'css', );
+mix.sass( `${devPath}/scss/screen.scss`, 'css', )
+	.sass( `${devPath}/scss/customize-preview.scss`,  'css' )
+	.sass( `${devPath}/scss/customize-controls.scss`, 'css' );